react: move backend call to projects list view
